applications/external/flipdeck/.host/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flipdeck-host"
version = "3.0.0"
description = "FlipDeck Host – USB HID macro pad companion for Flipper Zero"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pynput>=1.7",
    "flask>=3.0",
    "pyserial>=3.5",
    "pyyaml>=6.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
linux = ["evdev>=1.6"]

[project.scripts]
flipdeck = "flipdeck_host.cli:main"

[tool.setuptools.packages.find]
include = ["flipdeck_host*"]
